requirements.txt
# Core AI and Database
langchain==0.1.17
langchain-openai==0.1.7
langchain-community==0.0.38
pymongo==4.7.2
motor==3.4.0

# Google Cloud
google-cloud-functions-framework==3.5.0
google-cloud-secret-manager==2.18.3
google-cloud-monitoring==2.19.0

# LINE Bot
line-bot-sdk==3.9.0

# Utilities
python-dotenv==1.0.0
pydantic==2.7.1
asyncio==3.4.3

# Airtable integration
pyairtable==2.3.3
requests==2.31.0

# Development
pytest==8.2.1
pytest-asyncio==0.23.7
black==24.4.2